feat(cosmos/scripts): Allow applying patch of the cargo files during compilation
We faced problem deploying to `Cudos` due to dissonance with the verions of
some the `cosmwams` packages we use. That's why we introduce a way to
"patch" them, by overwring the `Cargo` files during the contract compilation.
This solved the problem for `Cudos` and potentionaly other networks we will
target in the future.

To compile the `Cudos` version of the verifier contract execute the folloing:
`ts-node /home/Emil/code/repos/metacraft-labs/DendrETH/contracts/cosmos/scripts/verifier-compile-contract-and-tools-script.ts run --patch cudos`