fix: Make SHA256 tests more explicit
The tests now reuse the code for the test logic. They differ only in the input
data and the expected messages.

The inputs for the tests are now explicitly specified as bytes in hex, as well
as the expected output. This surfaces that the last byte of the succeeding test
is 0x00, which might be not have been the intention, reading the original code.

Randomness is removed from the failing test, as it is an unnecessary
complication and reduces reproducibility of the test.

A new test is added that makes sure the SHA256 implementation corresponds to
that of coreutils-9.3 for one real ASCII string.